Create views for org/meeting creation
This commit adds forms and views to create an organisation or meeting,
including M2M fields, placing the creator of an org as manager and
for meetings as a president. Only managers can create a meeting.

Further, a list of meetings is a stub for the org home page. Mixins were
created for permissions checks.

Meeting creation will work but return an error as no success_url is
provided. The success URL will be created later, and the view should be
for the creation of the agenda.