Create Invitation models
These models, one for meetings and the other for organisations, will
allow managers to "invite" people who don't necessarily have an account
to join the site to participate in the role assigned to them. It uses
their email for contact, and includes individual keys to use to join the
meeting/org.